requirements.txt
asttokens==3.0.0
click==8.3.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
ipykernel==7.1.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.2
numpy==2.2.6
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.0
portalocker==3.2.0
prompt_toolkit==3.0.52
psutil==7.1.2
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
pyzmq==27.1.0
regex==2025.10.23
sacrebleu==2.5.1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
torch==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
wcwidth==0.2.14
